% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuctest.R
\name{f}
\alias{f}
\title{simple function}
\usage{
f(x)
}
\arguments{
\item{x}{x is a number}
}
\value{
a dataframe
}
\description{
today,I create my first function,a very usrful function.
}
\details{
This is some descriptio of this function.

you can use this function to caculate x+1,then return the value of x+1.
}
\examples{
x=1;f(x)
}
