% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmModFinal.R
\name{lmModFinal}
\alias{lmModFinal}
\title{Linear model}
\usage{
lmModFinal(formula, data, ...)
}
\arguments{
\item{formula}{formula to be fitted}

\item{data}{data}

\item{...}{other arguments}
}
\description{
This function estimates model parameters in a linear regression model
}
