% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{model_fit}
\usage{
model_fit(mf)
}
\arguments{
\item{mf}{a dataframe of complete data}
}
\value{
'model_fit' returns a list of results containing the following components:

y_hat: a n*1 matrix containing the fitted value

beta_hat: a n*1 matrix containing the fitted coefficients

residuals: a n*1 matrix containing the residuals (observed values - fitted values)
}
\description{
function to fit a linear model
}
\details{
you can use this function to fit for a linear model return a list of results.
}
\examples{
x = as.data.frame(matrix(c(2,4,6,8,1.1,1.9,3.1,4.2),4,2)); model_fit(x)
}
