% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  datasize1,
  datasize2,
  error_range,
  data_range,
  offsets_range,
  weights_range,
  coefficients
)
}
\arguments{
\item{datasize1}{a integer represents the amounts of every variables .}

\item{datasize2}{a integer represents the numbers of variables.}

\item{error_range}{the range of errors of outcomes. This should be NULL or a numeric vector of size datasize1}

\item{data_range}{a datasize2*2 matrix: every row represents the range of every variable.}

\item{offsets_range}{the range of expected offsets. This should be NULL or a numeric vector of size datasize1.}

\item{weights_range}{the range of expected weights. This should be NULL or a numeric vector of size datasize1.}

\item{coefficients}{a numeric vector with length datasize1: the simulated coefficient for every variable.}
}
\value{
'simulate' returns a dataframe of results containing the following components:

V1: the simulated Outcome. It should be a numeric vector with length of datasize1

V2, V3,...: the simulated Predictors. It should be a numeric vector with length of datasize1

(weights): If not NULL, the simulated weights. It should be a numeric vector with length of datasize1

(offsets): If not NULL, the simulated offsets. It should be a numeric vector with length of datasize1
}
\description{
function to generate simulation data for test
}
\details{
you can use this function to create random datasets for testing.
}
\examples{
data_range = matrix(c(0,0,0,1,1,1),3,2)
offsets_range = c(0,1)
weights_range = c(0,1)
coefficients = c(1,2,3)
data = simulate_data(10, 3, error_range= c(0,1), data_range, offsets_range, weights_range,coefficients)
}
