% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{model_fit}
\usage{
model_fit(mf, intercept, model = TRUE, x = FALSE, y = FALSE, qr = TRUE, offset)
}
\arguments{
\item{mf}{a dataframe of complete data}

\item{intercept}{an interger equals to 0 or 1. If 0, intercept should not be included in the model}

\item{model}{logical. If TRUE the the model frame is returned.}

\item{x}{logical. If TRUE the the model matrix is returned.}

\item{y}{logical. If TRUE the response is returned.}

\item{qr}{logical. If TRUE the the QR is returned.}

\item{offset}{'offset' is a vector with the same length of data. If null, the data frame "mf" does not contain an offset column, else, the last column is offset vector}
}
\value{
'model_fit' returns a list of results containing the following components:

fitted.values: a n*1 matrix containing the fitted value

beta_hat: a n*1 matrix containing the fitted coefficients

coefficients: a n*1 matrix containing the residuals (observed values - fitted values)

residuals: a n*1 matrix containing the residuals (observed values - fitted values)

offset: if not null, a n*1 matrix containing the offset values.
}
\description{
function to fit a linear model
}
\details{
you can use this function to fit for a linear model return a list of results.
}
\examples{
x = as.data.frame(matrix(c(2,4,6,8,1.1,1.9,3.1,4.2,1,1,1,1),4,3))
m = model_fit(x,1,TRUE, TRUE, TRUE, TRUE, offset = c(1,1,1,1))
}
